requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
discord.py==2.2.2
distlib==0.3.6
filelock==3.10.7
frozenlist==1.3.3
identify==2.5.22
idna==3.4
multidict==6.0.4
mutagen==1.46.0
nodeenv==1.7.0
platformdirs==3.2.0
pre-commit==3.2.1
pycparser==2.21
pycryptodomex==3.17
PyNaCl==1.5.0
python-dotenv==1.0.0
pytube==12.1.3
PyYAML==6.0
virtualenv==20.21.0
websockets==10.4
yarl==1.8.2
youtube-dl==2021.12.17
yt-dlp==2023.3.4
